Move find_egg_info to operations.generate_metadata
